@@ -9,7 +9,7 @@ namespace Easy
     {
         IUser CurrentUser { get; }
         IUser CurrentCustomer { get; }
-        IHostingEnvironment HostingEnvironment { get; }
+        IWebHostEnvironment HostingEnvironment { get; }
         bool IsAuthenticated { get; }
         T As<T>() where T : class, IApplicationContext;
         T Get<T>(string name);
