@@ -7,8 +7,8 @@ namespace Easy.Mvc.StateProviders
 {
     public class HostingEnvironmentStateProvider : IApplicationContextStateProvider
     {
-        private readonly IHostingEnvironment _hostingEnvironment;
-        public HostingEnvironmentStateProvider(IHostingEnvironment hostingEnvironment)
+        private readonly IWebHostEnvironment _hostingEnvironment;
+        public HostingEnvironmentStateProvider(IWebHostEnvironment hostingEnvironment)
         {
             _hostingEnvironment = hostingEnvironment;
         }
