@@ -5,13 +5,19 @@
  * http://www.zkea.net/licenses
  */
 using Easy;
+using Easy.Mvc.Plugin;
 using Easy.Mvc.Resource;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using System;
 using ZKEACMS.DbConnectionPool;
 
 namespace ZKEACMS.WebHost
@@ -27,10 +33,17 @@ namespace ZKEACMS.WebHost
         public void ConfigureServices(IServiceCollection services)
         {
             services.ConfigureResource<DefaultResourceManager>();
+            services.TryAddEnumerable(ServiceDescriptor.Singleton<IConfigureOptions<MvcRazorRuntimeCompilationOptions>, CompilationOptionsSetup>());
+            Type mvcBuilderType = typeof(Builder);
+            Easy.Mvc.Plugin.PluginActivtor.LoadedPlugins.Add(new PluginDescriptor
+            {
+                Assembly = mvcBuilderType.Assembly,
+                PluginType = mvcBuilderType
+            });
             services.UseZKEACMS(Configuration);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
         {
             if (env.IsDevelopment())
             {
@@ -38,7 +51,6 @@ namespace ZKEACMS.WebHost
             }
             else
             {
-               // loggerFactory.UseFileLog(env, app.ApplicationServices.GetService<IHttpContextAccessor>());
                 app.UseExceptionHandler("/Error");
                 app.UseStatusCodePagesWithReExecute("/Error/Code/{0}");
             }
