@@ -20,20 +20,14 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Session;
-using Microsoft.AspNetCore.Mvc.ApplicationParts;
-using Microsoft.AspNetCore.Mvc.Razor;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.DependencyModel;
-using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Hosting;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using ZKEACMS.Account;
 using ZKEACMS.Article.Models;
 using ZKEACMS.Common.Models;
@@ -54,10 +48,10 @@ using ZKEACMS.Route;
 using ZKEACMS.Setting;
 using ZKEACMS.SMTP;
 using ZKEACMS.Theme;
+using ZKEACMS.Validate;
 using ZKEACMS.Widget;
 using ZKEACMS.WidgetTemplate;
 using ZKEACMS.Zone;
-using ZKEACMS.Validate;
 
 namespace ZKEACMS
 {
@@ -76,11 +70,10 @@ namespace ZKEACMS
                 opt.Cookie.IsEssential = true;
             });
 
-            IMvcBuilder mvcBuilder = services.AddMvc(option =>
+            IMvcBuilder mvcBuilder = services.AddControllersWithViews(option =>
              {
                  option.ModelBinderProviders.Insert(0, new WidgetTypeModelBinderProvider());
                  option.ModelMetadataDetailsProviders.Add(new DataAnnotationsMetadataProvider());
-                 //option.EnableEndpointRouting = false;
              })
             .AddRazorOptions(opt =>
             {
@@ -88,9 +81,14 @@ namespace ZKEACMS
                 opt.ViewLocationExpanders.Add(new ThemeViewLocationExpander());
             })
             .AddControllersAsServices()
-            .AddJsonOptions(option => { option.SerializerSettings.DateFormatString = "yyyy-MM-dd"; })
+            .AddNewtonsoftJson(options =>
+            {
+                options.SerializerSettings.DateFormatString = "yyyy/MM/dd H:mm";
+            })
             .SetCompatibilityVersion(CompatibilityVersion.Latest);
 
+            services.AddRazorPages();
+            services.AddHealthChecks();
             services.TryAddSingleton<IHttpContextAccessor, HttpContextAccessor>();
 
             services.TryAddScoped<IApplicationContextAccessor, ApplicationContextAccessor>();
@@ -139,6 +137,8 @@ namespace ZKEACMS
             services.AddTransient<IEventViewerService, EventViewerService>();
 
             services.AddTransient<IStorage, WebStorage>();
+
+            services.ConfigureStateProvider<StateProvider.OuterChainPictureStateProvider>();
 
             services.ConfigureCache<IEnumerable<WidgetBase>>();
             services.ConfigureCache<IEnumerable<ZoneEntity>>();
@@ -244,9 +244,8 @@ namespace ZKEACMS
                 });
         }
 
-        public static void UseZKEACMS(this IApplicationBuilder applicationBuilder, IHostingEnvironment hostingEnvironment, IHttpContextAccessor httpContextAccessor)
+        public static void UseZKEACMS(this IApplicationBuilder applicationBuilder, IWebHostEnvironment hostingEnvironment, IHttpContextAccessor httpContextAccessor)
         {
-            applicationBuilder.UseAuthentication();
             if (hostingEnvironment.IsDevelopment())
             {
                 applicationBuilder.UsePluginStaticFile();
@@ -256,13 +255,26 @@ namespace ZKEACMS
             applicationBuilder.ConfigureResource();
             applicationBuilder.ConfigurePlugin(hostingEnvironment);
             applicationBuilder.UseSession();
-            applicationBuilder.UseMvc(routes =>
+            applicationBuilder.UseRouting();
+
+            applicationBuilder.UseAuthentication();
+            applicationBuilder.UseAuthorization();
+
+            applicationBuilder.UseEndpoints(endpoints =>
             {
                 applicationBuilder.ApplicationServices.GetService<IRouteProvider>().GetRoutes().OrderByDescending(route => route.Priority).Each(route =>
                 {
-                    routes.MapRoute(route.RouteName, route.Template, route.Defaults, route.Constraints, route.DataTokens);
+                    endpoints.MapControllerRoute(
+                        name: route.RouteName,
+                        pattern: route.Template,
+                        defaults: route.Defaults,
+                        constraints: route.Constraints,
+                        dataTokens: route.DataTokens);
                 });
+
+                endpoints.MapRazorPages();
             });
+
             foreach (IStartTask task in applicationBuilder.ApplicationServices.GetServices<IStartTask>())
             {
                 task.Excute();
