@@ -77,7 +77,7 @@ namespace ZKEACMS.Filter
                 string path = string.Format(Layouts.Theme, name);
                 string path2 = string.Format(Layouts.Theme2, name);
 
-                var env = filterContext.HttpContext.RequestServices.GetService<IHostingEnvironment>();
+                var env = filterContext.HttpContext.RequestServices.GetService<IWebHostEnvironment>();
                 var controller = filterContext.Controller as Controller;
 
                 if (File.Exists(env.MapPath(controller.Url.ToArray(path)))) return path;
