@@ -17,10 +17,18 @@ using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyModel;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
+using Microsoft.AspNetCore.Mvc.Razor.Compilation;
+using Microsoft.AspNetCore.Razor.Hosting;
+using Microsoft.Extensions.Primitives;
 
 namespace ZKEACMS
 {
-    public abstract class PluginBase : ResourceManager, IRouteRegister, IPluginStartup, IApplicationPartTypeProvider, ICompilationReferencesProvider
+    public abstract class PluginBase : ResourceManager,
+        IRouteRegister,
+        IPluginStartup,
+        IApplicationPartTypeProvider,
+        ICompilationReferencesProvider,
+        IApplicationFeatureProvider<ViewsFeature>
     {
         private const string ControllerTypeNameSuffix = "Controller";
         public Assembly Assembly { get; set; }
@@ -30,13 +38,17 @@ namespace ZKEACMS
         public abstract IEnumerable<PermissionDescriptor> RegistPermission();
         public abstract IEnumerable<WidgetTemplateEntity> WidgetServiceTypes();
         public abstract void ConfigureServices(IServiceCollection serviceCollection);
-        public virtual void ConfigureApplication(IApplicationBuilder app, IHostingEnvironment env)
+        public virtual void ConfigureApplication(IApplicationBuilder app, IWebHostEnvironment env)
         {
 
         }
         public virtual void ConfigureMVC(IMvcBuilder mvcBuilder)
         {
-            mvcBuilder.ConfigureApplicationPartManager(manguage => manguage.ApplicationParts.Add(this));
+            mvcBuilder.ConfigureApplicationPartManager(manguage =>
+            {
+                manguage.FeatureProviders.Add(this);
+                manguage.ApplicationParts.Add(this);
+            });
         }
         static Dictionary<Type, string> pluginPathCache = new Dictionary<Type, string>();
         static Dictionary<Type, string> pluginNameCache = new Dictionary<Type, string>();
@@ -86,7 +98,7 @@ namespace ZKEACMS
                     item.PluginName = this.Name;
                     AdminMenus.Menus.Add(item);
                 }
-                
+
             }
             this.SetupResource();
             var permissions = this.RegistPermission();
@@ -96,7 +108,7 @@ namespace ZKEACMS
                 {
                     item.PluginName = this.Name;
                     PermissionKeys.KnownPermissions.Add(item);
-                }                
+                }
             }
             var routes = this.RegistRoute();
             if (routes != null)
@@ -195,5 +207,58 @@ namespace ZKEACMS
 
             return new[] { Assembly.Location };
         }
+        #region Viewfeature
+        public virtual void PopulateFeature(IEnumerable<ApplicationPart> parts, ViewsFeature feature)
+        {
+            var knownIdentifiers = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+            var attributes = GetViewAttributesLegacy(Assembly);
+            foreach (var item in attributes)
+            {
+                var descriptor = new CompiledViewDescriptor(item);
+                if (knownIdentifiers.Add(descriptor.RelativePath))
+                {
+                    feature.ViewDescriptors.Add(descriptor);
+                }
+            }
+        }
+
+
+        protected virtual IEnumerable<RazorCompiledItem> GetViewAttributesLegacy(Assembly assembly)
+        {
+            if (assembly == null)
+            {
+                throw new ArgumentNullException(nameof(assembly));
+            }
+
+            var featureAssembly = GetViewAssembly(assembly);
+            if (featureAssembly != null)
+            {
+                return new RazorCompiledItemLoader().LoadItems(featureAssembly);
+            }
+
+            return Enumerable.Empty<RazorCompiledItem>();
+        }
+
+        protected virtual Assembly GetViewAssembly(Assembly assembly)
+        {
+            if (assembly.IsDynamic || string.IsNullOrEmpty(assembly.Location))
+            {
+                return null;
+            }
+            string[] viewAssemblySuffixes = new string[] { ".Views", ".PrecompiledViews" };
+            for (var i = 0; i < viewAssemblySuffixes.Length; i++)
+            {
+                var fileName = $"{assembly.GetName().Name}{viewAssemblySuffixes[i]}.dll";
+                var filePath = Path.Combine(Path.GetDirectoryName(assembly.Location), fileName);
+
+                if (File.Exists(filePath))
+                {
+                    return Assembly.LoadFile(filePath);
+                }
+            }
+
+            return null;
+        }
+        #endregion
     }
 }
