@@ -16,7 +16,7 @@ namespace ZKEACMS.Widget
 {
     public class WidgetPackageInstaller : FilePackageInstaller
     {
-        public WidgetPackageInstaller(IHostingEnvironment hostingEnvironment) : base(hostingEnvironment)
+        public WidgetPackageInstaller(IWebHostEnvironment hostingEnvironment) : base(hostingEnvironment)
         {
         }
 
@@ -66,7 +66,7 @@ namespace ZKEACMS.Widget
     public class DataDictionaryPackageInstaller : FilePackageInstaller
     {
         private readonly IDataDictionaryService _dataDictionaryService;
-        public DataDictionaryPackageInstaller(IHostingEnvironment hostingEnvironment, IDataDictionaryService dataDictionaryService) : base(hostingEnvironment)
+        public DataDictionaryPackageInstaller(IWebHostEnvironment hostingEnvironment, IDataDictionaryService dataDictionaryService) : base(hostingEnvironment)
         {
             _dataDictionaryService = dataDictionaryService;
         }
