@@ -50,10 +50,10 @@ namespace Easy.Mvc.Controllers
                     query.ValueMin = Easy.Reflection.ClassAction.ValueConvert(p, item.Search.ValueMin);
                     query.ValueMax = Easy.Reflection.ClassAction.ValueConvert(p, item.Search.ValueMax);
 
-                    if (query.ValueMax != null && query.ValueMax is DateTime)
-                    {
-                        query.ValueMax = ((DateTime)query.ValueMax).AddDays(1).AddMilliseconds(-1);
-                    }
+                    //if (query.ValueMax != null && query.ValueMax is DateTime)
+                    //{
+                    //    query.ValueMax = ((DateTime)query.ValueMax).AddDays(1).AddMilliseconds(-1);
+                    //}
                 }
                 catch
                 {
