@@ -8,21 +8,28 @@ using ZKEACMS.Widget;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 using Easy.Extend;
+using System.Collections.Concurrent;
 
 namespace ZKEACMS.Product.Service
 {
     public class ProductDetailWidgetService : WidgetService<ProductDetailWidget>
     {
         private const string ProductDetailWidgetRelatedPageUrls = "ProductDetailWidgetRelatedPageUrls";
+        private readonly ConcurrentDictionary<string, object> _allRelatedUrlCache;
         private readonly IProductService _productService;
-        public ProductDetailWidgetService(IWidgetBasePartService widgetService, IProductService productService, IApplicationContext applicationContext, CMSDbContext dbContext)
+        public ProductDetailWidgetService(IWidgetBasePartService widgetService,
+            IProductService productService,
+            IApplicationContext applicationContext,
+            Easy.Cache.ICacheManager<ConcurrentDictionary<string, object>> cacheManager,
+            CMSDbContext dbContext)
             : base(widgetService, applicationContext, dbContext)
         {
             _productService = productService;
+            _allRelatedUrlCache = cacheManager.GetOrAdd(ProductDetailWidgetRelatedPageUrls, new ConcurrentDictionary<string, object>());
         }
         private void DismissRelatedPageUrls()
         {
-            ProductPlug.AllRelatedUrlCache.TryRemove(ProductDetailWidgetRelatedPageUrls, out var urls);
+            _allRelatedUrlCache.TryRemove(ProductDetailWidgetRelatedPageUrls, out var urls);
         }
 
         public override void AddWidget(WidgetBase widget)
@@ -76,11 +83,11 @@ namespace ZKEACMS.Product.Service
 
         public string[] GetRelatedPageUrls()
         {
-            return ProductPlug.AllRelatedUrlCache.GetOrAdd(ProductDetailWidgetRelatedPageUrls, fac =>
+            return _allRelatedUrlCache.GetOrAdd(ProductDetailWidgetRelatedPageUrls, fac =>
             {
                 var pages = WidgetBasePartService.Get(w => Get().Select(m => m.ID).Contains(w.ID)).Select(m => m.PageID).ToArray();
                 return DbContext.Page.Where(p => pages.Contains(p.ID)).Select(m => m.Url.Replace("~/", "/")).Distinct().ToArray();
-            });
+            }) as string[];
         }
     }
 }
