@@ -71,8 +71,8 @@ namespace ZKEACMS.WebHost
                 .Include($"{ScriptPath}/dataTable.js", $"{ScriptPath}/dataTable.min.js");
 
             script("datepicker")
-                .Include($"{LibraryPath}/bootstrap-datepicker/dist/js/bootstrap-datepicker.js", $"{LibraryPath}/bootstrap-datepicker/dist/js/bootstrap-datepicker.min.js")
-                .Include($"{LibraryPath}/bootstrap-datepicker/dist/locales/bootstrap-datepicker.zh-CN.min.js");
+                .Include($"{LibraryPath}/bootstrap-datepicker/moment-with-locales.min.js")
+                .Include($"{LibraryPath}/bootstrap-datepicker/bootstrap-datetimepicker.min.js");
 
             script("pjax")
                 .Include($"{LibraryPath}/jquery-pjax/jquery.pjax.js");
@@ -94,7 +94,8 @@ namespace ZKEACMS.WebHost
                 .Include($"{LibraryPath}/codemirror/mode/xml/xml.js")
                 .Include($"{LibraryPath}/codemirror/mode/javascript/javascript.js")
                 .Include($"{LibraryPath}/codemirror/mode/css/css.js")
-                .Include($"{LibraryPath}/codemirror/mode/htmlmixed/htmlmixed.js");
+                .Include($"{LibraryPath}/codemirror/mode/htmlmixed/htmlmixed.js")
+                .Include($"{LibraryPath}/codemirror/addon/format/autoformat.js", $"{LibraryPath}/codemirror/addon/format/autoformat.min.js");
         }
 
         protected override void InitStyle(Func<string, ResourceHelper> style)
@@ -139,7 +140,7 @@ namespace ZKEACMS.WebHost
                 .Include($"{LibraryPath}/datatables/media/css/dataTables.bootstrap.css", $"{LibraryPath}/datatables/media/css/dataTables.bootstrap.min.css");
 
             style("datepicker")
-                .Include($"{LibraryPath}/bootstrap-datepicker/dist/css/bootstrap-datepicker3.css", $"{LibraryPath}/bootstrap-datepicker/dist/css/bootstrap-datepicker3.min.css");
+                .Include($"{LibraryPath}/bootstrap-datepicker/bootstrap-datetimepicker.min.css");
 
             style("select2")
                 .Include($"{LibraryPath}/select2/dist/css/select2.css", $"{LibraryPath}/select2/dist/css/select2.min.css");
