@@ -7,11 +7,11 @@
     }
 }
 
-<button type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+<button type="button" class="zk-btn zk-btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
     <span class="caret"></span>
-    <span class="sr-only">Toggle Dropdown</span>
+    <span class="sr-only"></span>
 </button>
-<ul class="dropdown-menu">
+<ul class="zk-dropdown-menu pull-right">
     <li>
         <a href="@Url.Action("Index", "Dashboard")">
             <i class="glyphicon glyphicon-dashboard"></i>
