@@ -1,6 +1,6 @@
 @using Microsoft.AspNetCore.Hosting
+@using Microsoft.Extensions.Hosting
 @model IEnumerable<ThemeEntity>
-@inject IHostingEnvironment hostingEnvironment
 <div class="panel panel-default">
     <div class="panel-heading">
         @L("主题")
@@ -20,7 +20,7 @@
                                 <label>
                                     <input type="radio" name="theme" @(item.IsActived ? "checked" : "") value="@item.ID"> @item.Title
                                 </label>
-                                @if (hostingEnvironment.IsDevelopment())
+                                @if (ApplicationContext.HostingEnvironment.IsDevelopment())
                                 {
                                     <div class="pull-right">
                                         <a class="download pack-theme" href="@Url.Action("ThemePackage", new {id = item.ID})" title="@L("打包")">
@@ -45,12 +45,12 @@
             <li>
                 @L("若主题有更新，重复安装即可更新")
             </li>
-            <li>
-                <span style="color:red;">友情提示: 切换主题时,会先删除大部分数据,然后导入主题自带数据,操作不可逆,请先线下备份数据库</span><br />
-                <span>删除主题数据的脚本位置: /wwwroot/themes/主题名称/sql/uninstall/*.sql</span><br />
-                <span>导入主题数据的脚本位置: /wwwroot/themes/主题名称/sql/install/*.sql</span><br />
-                <span>文件夹下脚本顺序很重要,切莫调整顺序</span>
-            </li>
+            @*<li>
+                    <span style="color:red;">友情提示: 切换主题时,会先删除大部分数据,然后导入主题自带数据,操作不可逆,请先线下备份数据库</span><br />
+                    <span>删除主题数据的脚本位置: /wwwroot/themes/主题名称/sql/uninstall/*.sql</span><br />
+                    <span>导入主题数据的脚本位置: /wwwroot/themes/主题名称/sql/install/*.sql</span><br />
+                    <span>文件夹下脚本顺序很重要,切莫调整顺序</span>
+                </li>*@
         </ul>
     </div>
     @if (Authorizer.Authorize(PermissionKeys.ManageTheme))
