@@ -118,9 +118,10 @@
                             }
                         }
                         if (columnSetting.dataType == "DateTime") {
-                            searchInput.find(".form-control").datepicker({
-                                language: "zh-CN",
-                                format: "yyyy/mm/dd" //columnSetting.format
+                            debugger
+                            searchInput.find(".form-control").datetimepicker({
+                                locale: 'zh-cn',
+                                format: columnSetting.format //columnSetting.format
                             });
                         }
                         searchInput.appendTo($(column.footer()));
