@@ -33,7 +33,7 @@ namespace ZKEACMS.Article.Models
     {
         protected override void ViewConfigure()
         {
-            ViewConfig(m => m.ID).AsHidden();
+            ViewConfig(m => m.ID).AsHidden().ShowInGrid().Order(1);
             ViewConfig(m => m.Title).AsTextBox().Required().Order(2).ShowInGrid().Search(Easy.LINQ.Query.Operators.Contains);
             ViewConfig(m => m.Url).AsTextBox().Order(3).MaxLength(100).UrlPart();
             ViewConfig(m => m.Status).AsDropDownList().DataSource(DicKeys.RecordStatus, SourceType.Dictionary);
